fixed overflow, (use long int)

+++ format-printer.c
@@ -13,7 +13,7 @@ int format_printer(const char *k, int *ind, va_list arg)
 {
 	int counter = 0, stray_len = 0;
 	char *s;
-	int c, i, d, u, b, o, x, X;
+	long int c, i, d, u, b, o, x, X;
 
 	switch (k[*ind])
 	{
@@ -36,35 +36,35 @@ int format_printer(const char *k, int *ind, va_list arg)
 
 		break;
 	case 'i':
-		i = va_arg(arg, int);
+		i = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(i, 10));
 		break;
 
 	case 'd':
-		d = va_arg(arg, int);
+		d = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(d, 10));
 		break;
 	case 'b':
-		b = va_arg(arg, int);
+		b = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(b, 2));
 		break;
 		case '%':
 		counter += _putchar('%');
 		break;
 	case 'u':
-		u = va_arg(arg, int);
+		u = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(u, 'u'));
 		break;
 	case 'o':
-		o = va_arg(arg, int);
+		o = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(o, 8));
 		break;
 	case 'x':
-		x = va_arg(arg, int);
+		x = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(x, 'x'));
 		break;
 	case 'X':
-		X = va_arg(arg, int);
+		X = va_arg(arg, long int);
 		counter += arrayprinter(base_converter(X, 16));
 		break;
 	case '\0':
