+++ main.h
@@ -9,7 +9,7 @@
 
 int _printf(const char *format, ...);
 int _putchar(char c);
-char *base_converter(int num, int base_n);
+char *base_converter(long int num, int base_n);
 int arrayprinter(char *array);
 int format_printer(const char *, int *, va_list arg);
